refactor(lang-core): organize telemetry modules
