+++ packages/lang-core/src/postinstall.ts
@@ -1,4 +1,4 @@
-import { runInstallTelemetry } from "./install-telemetry";
+import { runInstallTelemetry } from "./telemetry/install";
 
 void runInstallTelemetry().catch(() => {
   // The lifecycle entry must always resolve successfully.
